telegram-organizer-bot/requirements.txt
telethon>=1.36.0
httpx>=0.27.0
python-dotenv>=1.0.1
